fix: resolve all ruff lint errors blocking CI
- Add UP045 and UP017 to ruff ignore list (Optional[X] style + timezone alias)
- Replace int(round(...)) with round(...) — RUF046, already returns int
- Remove unused pytest imports from test files — F401
- Prefix unused unpacked variables with _ in tests — RUF059
- Replace EN DASH with hyphen in docstrings — RUF002

+++ analytics/providers/yfinance_provider.py
@@ -44,7 +44,7 @@ def _pct(v: Any) -> Optional[float]:
 
 def _safe_int(v: Any) -> Optional[int]:
     f = _safe(v)
-    return int(round(f)) if f is not None else None
+    return round(f) if f is not None else None
 
 
 def _infer_market(ticker: str) -> Market:
